+++ compose.yml
@@ -21,7 +21,6 @@ x-service-template: &service-template
     - node_modules_ghost_members-csv:/home/ghost/ghost/members-csv/node_modules:delegated
     - node_modules_ghost_mw-error-handler:/home/ghost/ghost/mw-error-handler/node_modules:delegated
     - node_modules_ghost_mw-vhost:/home/ghost/ghost/mw-vhost/node_modules:delegated
-    - node_modules_ghost_post-events:/home/ghost/ghost/post-events/node_modules:delegated
     - node_modules_ghost_prometheus-metrics:/home/ghost/ghost/prometheus-metrics/node_modules:delegated
     - node_modules_ghost_security:/home/ghost/ghost/security/node_modules:delegated
     - node_modules_apps_admin-x-activitypub:/home/ghost/apps/admin-x-activitypub/node_modules:delegated
@@ -184,7 +183,6 @@ volumes:
   node_modules_ghost_members-csv: {}
   node_modules_ghost_mw-error-handler: {}
   node_modules_ghost_mw-vhost: {}
-  node_modules_ghost_post-events: {}
   node_modules_ghost_prometheus-metrics: {}
   node_modules_ghost_security: {}
   node_modules_apps_admin-x-activitypub: {}
