+++ ghost/core/core/server/services/Users.js
@@ -6,7 +6,6 @@ const ObjectId = require('bson-objectid').default;
  * @TODO: pass these in as dependencies
  */
 const DomainEvents = require('@tryghost/domain-events/lib/DomainEvents');
-const {PostsBulkAddTagsEvent} = require('@tryghost/post-events');
 
 /**
  * @typedef {Object} IdbBackup
@@ -145,6 +144,7 @@ class Users {
         await this.models.Post.addActions('edited', usersPostIds, {transacting, context});
 
         // dispatch event to ensure collections are updated
+        const {PostsBulkAddTagsEvent} = require('../../shared/events-ts');
         DomainEvents.dispatch(PostsBulkAddTagsEvent.create(usersPostIds));
     }
