+++ ghost/core/core/shared/events-ts/PostDeletedEvent.ts
