+++ ghost/core/core/shared/events-ts/PostsBulkAddTagsEvent.ts
