+++ ghost/core/core/shared/events-ts/PostsBulkDestroyedEvent.ts
