+++ ghost/core/core/shared/events-ts/PostsBulkFeaturedEvent.ts
