+++ ghost/core/core/shared/events-ts/PostsBulkUnpublishedEvent.ts
