--- ghost/post-events/tsconfig.json
@@ -1,9 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "include": [
-        "src/**/*"
-    ],
-    "compilerOptions": {
-        "outDir": "build"
-    }
-}
